refactor: move test utils to common location
