+++ pkg/pdp/aggregator/pieceaggregator_test.go
@@ -3,16 +3,12 @@ package aggregator_test
 import (
 	"context"
 	"fmt"
-	"io"
-	"math/rand"
 	"testing"
-	"time"
 
-	commp "github.com/filecoin-project/go-fil-commp-hashhash"
 	"github.com/ipld/go-ipld-prime/datamodel"
-	"github.com/storacha/go-libstoracha/piece/digest"
 	"github.com/storacha/go-libstoracha/piece/piece"
 	"github.com/storacha/storage/internal/mocks"
+	"github.com/storacha/storage/pkg/internal/testutil"
 	"github.com/storacha/storage/pkg/pdp/aggregator"
 	"github.com/storacha/storage/pkg/pdp/aggregator/aggregate"
 	"github.com/storacha/storage/pkg/pdp/aggregator/fns"
@@ -40,6 +36,10 @@ func setupPieceAggregatorDependencies(
 	return workspaceMock, storeMock, queueSubmissionMock, baMock
 }
 
+const (
+	MB = 1 << 20
+)
+
 func TestPieceAggregator_StoreAndSubmit(t *testing.T) {
 	ctx := context.Background()
 	ctrl := gomock.NewController(t)
@@ -50,22 +50,12 @@ func TestPieceAggregator_StoreAndSubmit(t *testing.T) {
 
 	pa := aggregator.NewPieceAggregator(workspaceMock, storeMock, queueSubMock, aggregator.WithAggregator(baMock))
 	// the below makes assertion that when three aggregates are returned by the aggregator of the piece-aggregator
-	// three writes are made to the aggregate-store and three submissions are made too the queue-submission function.
-	p1 := createPiece(t, MB)
-	p2 := createPiece(t, MB)
-	p3 := createPiece(t, MB)
+	// three writes are made to the aggregate-store and three submissions are made to the queue-submission function.
+	p1 := testutil.CreatePiece(t, MB)
+	p2 := testutil.CreatePiece(t, MB)
+	p3 := testutil.CreatePiece(t, MB)
 	expectedPieces := []piece.PieceLink{p1, p2, p3}
-	expectedAggregates := []aggregate.Aggregate{
-		{
-			Root: p1,
-		},
-		{
-			Root: p2,
-		},
-		{
-			Root: p3,
-		},
-	}
+	expectedAggregates := []aggregate.Aggregate{{Root: p1}, {Root: p2}, {Root: p3}}
 	expectedBuffer := fns.Buffer{}
 
 	workspaceMock.EXPECT().GetBuffer(ctx).Return(expectedBuffer, nil)
@@ -90,15 +80,9 @@ func TestPieceAggregator_GetBufferError(t *testing.T) {
 	workspaceMock, storeMock, queueSubMock, baMock := setupPieceAggregatorDependencies(ctrl, &submittedLinks)
 
 	pa := aggregator.NewPieceAggregator(workspaceMock, storeMock, queueSubMock, aggregator.WithAggregator(baMock))
-	p1 := createPiece(t, MB)
-	p2 := createPiece(t, MB)
-	p3 := createPiece(t, MB)
-	expectedPieces := []piece.PieceLink{p1, p2, p3}
-	expectedBuffer := fns.Buffer{}
+	workspaceMock.EXPECT().GetBuffer(ctx).Return(fns.Buffer{}, fmt.Errorf("get buffer error"))
 
-	workspaceMock.EXPECT().GetBuffer(ctx).Return(expectedBuffer, fmt.Errorf("get buffer error"))
-
-	err := pa.AggregatePieces(ctx, expectedPieces)
+	err := pa.AggregatePieces(ctx, nil)
 	require.Error(t, err)
 
 	require.Len(t, submittedLinks, 0)
@@ -113,23 +97,11 @@ func TestPieceAggregator_PutBufferError(t *testing.T) {
 	workspaceMock, storeMock, queueSubMock, baMock := setupPieceAggregatorDependencies(ctrl, &submittedLinks)
 
 	pa := aggregator.NewPieceAggregator(workspaceMock, storeMock, queueSubMock, aggregator.WithAggregator(baMock))
-	// the below makes assertion that when three aggregates are returned by the aggregator of the piece-aggregator
-	// three writes are made to the aggregate-store and three submissions are made too the queue-submission function.
-	p1 := createPiece(t, MB)
-	p2 := createPiece(t, MB)
-	p3 := createPiece(t, MB)
+	p1 := testutil.CreatePiece(t, MB)
+	p2 := testutil.CreatePiece(t, MB)
+	p3 := testutil.CreatePiece(t, MB)
 	expectedPieces := []piece.PieceLink{p1, p2, p3}
-	expectedAggregates := []aggregate.Aggregate{
-		{
-			Root: p1,
-		},
-		{
-			Root: p2,
-		},
-		{
-			Root: p3,
-		},
-	}
+	expectedAggregates := []aggregate.Aggregate{{Root: p1}, {Root: p2}, {Root: p3}}
 	expectedBuffer := fns.Buffer{}
 
 	workspaceMock.EXPECT().GetBuffer(ctx).Return(expectedBuffer, nil)
@@ -151,11 +123,9 @@ func TestPieceAggregator_AggregatePieceError(t *testing.T) {
 	workspaceMock, storeMock, queueSubMock, baMock := setupPieceAggregatorDependencies(ctrl, &submittedLinks)
 
 	pa := aggregator.NewPieceAggregator(workspaceMock, storeMock, queueSubMock, aggregator.WithAggregator(baMock))
-	// the below makes assertion that when three aggregates are returned by the aggregator of the piece-aggregator
-	// three writes are made to the aggregate-store and three submissions are made too the queue-submission function.
-	p1 := createPiece(t, MB)
-	p2 := createPiece(t, MB)
-	p3 := createPiece(t, MB)
+	p1 := testutil.CreatePiece(t, MB)
+	p2 := testutil.CreatePiece(t, MB)
+	p3 := testutil.CreatePiece(t, MB)
 	expectedPieces := []piece.PieceLink{p1, p2, p3}
 	expectedBuffer := fns.Buffer{}
 
@@ -168,41 +138,29 @@ func TestPieceAggregator_AggregatePieceError(t *testing.T) {
 	require.Len(t, submittedLinks, 0)
 }
 
-const (
-	MB = 1 << 20
-)
+func TestPieceAggregator_StorePutError(t *testing.T) {
+	ctx := context.Background()
+	ctrl := gomock.NewController(t)
+	defer ctrl.Finish()
 
-// TODO(forrest): move this to a shared location with nfs pkg.
+	var submittedLinks []datamodel.Link
+	workspaceMock, storeMock, queueSubMock, baMock := setupPieceAggregatorDependencies(ctrl, &submittedLinks)
 
-// createPiece is a helper that produces a piece with the given unpadded size,
-// using random data so we don't rely on any pre-computed fixtures.
-func createPiece(t *testing.T, unpaddedSize int64) piece.PieceLink {
-	t.Helper()
+	pa := aggregator.NewPieceAggregator(workspaceMock, storeMock, queueSubMock, aggregator.WithAggregator(baMock))
+	p1 := testutil.CreatePiece(t, MB)
+	p2 := testutil.CreatePiece(t, MB)
+	p3 := testutil.CreatePiece(t, MB)
+	expectedPieces := []piece.PieceLink{p1, p2, p3}
+	expectedBuffer := fns.Buffer{}
+	expectedAggregates := []aggregate.Aggregate{{Root: p1}, {Root: p2}, {Root: p3}}
 
-	r := rand.New(rand.NewSource(time.Now().UnixNano()))
-	dataReader := io.LimitReader(r, unpaddedSize)
+	workspaceMock.EXPECT().GetBuffer(ctx).Return(expectedBuffer, nil)
+	baMock.EXPECT().AggregatePieces(expectedBuffer, expectedPieces).Return(expectedBuffer, expectedAggregates, nil)
+	workspaceMock.EXPECT().PutBuffer(ctx, expectedBuffer).Return(nil)
+	storeMock.EXPECT().Put(ctx, expectedAggregates[0].Root.Link(), expectedAggregates[0]).Return(fmt.Errorf("put buffer error"))
 
-	calc := &commp.Calc{}
-	n, err := io.Copy(calc, dataReader)
-	require.NoError(t, err, "failed copying data into commp.Calc")
-	require.Equal(t, unpaddedSize, n)
+	err := pa.AggregatePieces(ctx, expectedPieces)
+	require.Error(t, err)
 
-	commP, paddedSize, err := calc.Digest()
-	require.NoError(t, err, "failed to compute commP")
-
-	pieceDigest, err := digest.FromCommitmentAndSize(commP, uint64(unpaddedSize))
-	require.NoError(t, err, "failed building piece digest from commP")
-
-	p := piece.FromPieceDigest(pieceDigest)
-	// Ensure our piece’s PaddedSize matches the commp library’s reported paddedSize.
-	require.Equal(t, paddedSize, p.PaddedSize())
-
-	t.Logf("Created test piece: %s from unpadded size: %d", pieceLinkString(p), unpaddedSize)
-	return p
-}
-
-// pieceLinkString is a helper to display piece metadata in logs.
-func pieceLinkString(p piece.PieceLink) string {
-	return fmt.Sprintf("Piece: %s, Height: %d, Padding: %d, PaddedSize: %d",
-		p.Link(), p.Height(), p.Padding(), p.PaddedSize())
+	require.Len(t, submittedLinks, 0)
 }
